requirements.txt
loguru==0.7.3
langchain-openai==0.3.33
langchain-core==0.3.76
python-dotenv==1.0.1
sentence-transformers==5.1.0
numpy==2.3.3
streamlit==1.49.1